chore: test cases + fixes for streaming tool call (#3031)
